Extract helper module to separate file
And properly include it.
